Added Import Wizard Template, View Peers (Tree)
The wizard is just a quick template for some additional settings I plan
to add. I also added some GUI viewing functionality for peers via a
JTree. The coding for displaying is pretty messy (more or less builds
the entire thing in one go), but I don't intend to have partial updates
in the future.