Added PeerCreator, LoaderGUI
Made some changes besides that. For example, DataLoader now accepts
Files instead of strings to files. I updated some stuff so that it could
process and accept multiple files at once. PeerCreator is still lacking
in some regards, and I've temporarily stopped working on the PeerView
tree. I will try to get to that, but the creator is doing fairly well in
that function. I need to add more functions to the creator, such as bulk
creation. I also want to add a couple more variables to the database
format eventually. I'm tempted to plan/start the visuals, but I will
have to keep up progress on the objects and foundation code first.