Added application-wide data updates with changes
The code I added allows for windows to update as data is changed. I'm
not sure this is the right way to do it, but it is a thousand times
simpler than having individual listeners for each window I create. The
update mechanism is strictly within the GUI - the Data class doesn't do
anything, so I guess there is still a bit of a barrier between data and
view. I just got back into this after days of looking for
inspiration/motivation. Of course, there is still a bit of code to be
polished on what I added. I will likely clean stuff up tomorrow.